Process conflict.

+++ tiseg/models/segmentors/nuclei_cdnet.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
@@ -267,8 +268,8 @@ class NucleiCDNet(BaseSegmentor):
         """Simple test with single image."""
         seg_logit = self.inference(img, meta, rescale)
         seg_pred = seg_logit.argmax(dim=1)
-        # Extract nuclei inside class
-        seg_pred = (seg_pred == 1).long()
+        # Extract inside class
+        seg_pred = (seg_pred == 1)
         seg_pred = seg_pred.cpu().numpy()
         # unravel batch dim
         seg_pred = list(seg_pred)
@@ -289,7 +290,7 @@ class NucleiCDNet(BaseSegmentor):
         seg_logit /= len(imgs)
         seg_pred = seg_logit.argmax(dim=1)
         # Extract inside class
-        seg_pred = (seg_pred == 1).long()
+        seg_pred = (seg_pred == 1)
         seg_pred = seg_pred.cpu().numpy()
         # unravel batch dim
         seg_pred = list(seg_pred)
